FindICU.cmake changes to support more systems and more configurations and additional changes to allow linking against static library versions of ICU.
The existing ICU infrastructure was insufficient for multiple things.
- It relied on pkg-config, which is problematic on systems that don't
ship pkg-config (e.g. Ubuntu 12.04, Windows?)
- It didn't allow an easy way for specifying alternative locations for ICU
which is needed if you need to supply a non-system installed version.
This form is also useful when you want to supply a static library
version.
- It didn't handle linking when ICU was a static library. (libicudata is
	needed)
- pkg-config also doesn't handle libicudata
- pkg-config also can present problems for cross-compiling. (Android was
	also taken into account for this patch.)

This change fixes all these problems so Swift can be built on systems
like SteamOS/Steam-Runtime which don't supply libICU (and is based on
Ubuntu 12.04 which lacks a pkg-config). This allows you to supply
some other libICU which is a static library, and the Swift build system
will do the right thing.

An explicit setting of ICU_ROOT (cmake -DICU_ROOT=/path/to/icu_root) or
setting an environmental variable called ICU_ROOT will allow alternative
paths to be searched.